WIP. Tests are running.
Well and that's almost all about it. The security is a mere skeleton without functionality. I have to think about how I want to access the data - I think I want to use the services but they depend on JPA and DI of the JEE container and I don't want to rely on them in a JASPIC module. So I need to find another solution I like.